requirements.txt
fastapi
uvicorn
pillow-heif
Pillow
typing
